refactoring Notice

+++ DAQSystem/DAQSystem/CupMiniTCB.hh
@@ -19,103 +19,103 @@ public:
   void TriggerStart() override;
   void TriggerStop() override;
 
-  unsigned long ReadBCount(unsigned long mid) override { return 0; }
-  int ReadData(unsigned long mid, unsigned long bcount, unsigned char * data) override { return 0; }
+  uint32_t ReadBCount(uint32_t mid) override { return 0; }
+  int ReadData(uint32_t mid, uint32_t bcount, unsigned char * data) override { return 0; }
 
-  void WriteCW(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadCW(unsigned long mid, unsigned long ch) override;
-  void WriteGW(unsigned long mid, unsigned long data) override;
-  unsigned long ReadGW(unsigned long mid) override;
-  void WriteRL(unsigned long mid, unsigned long data) override;
-  unsigned long ReadRL(unsigned long mid) override;
-  void WriteDRAMON(unsigned long mid, unsigned long data) override;
-  unsigned long ReadDRAMON(unsigned long mid) override;
-  void WriteDACOFF(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadDACOFF(unsigned long mid, unsigned long ch) override;
-  void MeasurePED(unsigned long mid, unsigned long ch) override;
-  unsigned long ReadPED(unsigned long mid, unsigned long ch) override;
-  void WriteDLY(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadDLY(unsigned long mid, unsigned long ch) override;
-  void WriteTHR(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadTHR(unsigned long mid, unsigned long ch) override;
-  void WritePOL(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadPOL(unsigned long mid, unsigned long ch) override;
-  void WritePSW(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadPSW(unsigned long mid, unsigned long ch) override;
-  void WriteAMODE(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadAMODE(unsigned long mid, unsigned long ch) override;
-  void WritePCT(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadPCT(unsigned long mid, unsigned long ch) override;
-  void WritePCI(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadPCI(unsigned long mid, unsigned long ch) override;
-  void WritePWT(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadPWT(unsigned long mid, unsigned long ch) override;
-  void WriteDT(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadDT(unsigned long mid, unsigned long ch) override;
-  void WritePSS(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadPSS(unsigned long mid, unsigned long ch) override;
-  void WriteTM(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadTM(unsigned long mid, unsigned long ch) override;
-  void WriteTLT(unsigned long mid, unsigned long data) override;
-  unsigned long ReadTLT(unsigned long mid) override;
-  void WriteSTLT(unsigned long mid, unsigned long ch, unsigned long data) override;
-  unsigned long ReadSTLT(unsigned long mid, unsigned long ch) override;
-  void WriteDSR(unsigned long mid, unsigned long data) override;
-  unsigned long ReadDSR(unsigned long mid) override;
-  void AlignFADC(unsigned long mid) override;
-  void AlignSADC(unsigned long mid) override;
-  void AlignGADC(unsigned long mid) override;
+  void WriteCW(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadCW(uint32_t mid, uint32_t ch) override;
+  void WriteGW(uint32_t mid, uint32_t data) override;
+  uint32_t ReadGW(uint32_t mid) override;
+  void WriteRL(uint32_t mid, uint32_t data) override;
+  uint32_t ReadRL(uint32_t mid) override;
+  void WriteDRAMON(uint32_t mid, uint32_t data) override;
+  uint32_t ReadDRAMON(uint32_t mid) override;
+  void WriteDACOFF(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadDACOFF(uint32_t mid, uint32_t ch) override;
+  void MeasurePED(uint32_t mid, uint32_t ch) override;
+  uint32_t ReadPED(uint32_t mid, uint32_t ch) override;
+  void WriteDLY(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadDLY(uint32_t mid, uint32_t ch) override;
+  void WriteTHR(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadTHR(uint32_t mid, uint32_t ch) override;
+  void WritePOL(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadPOL(uint32_t mid, uint32_t ch) override;
+  void WritePSW(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadPSW(uint32_t mid, uint32_t ch) override;
+  void WriteAMODE(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadAMODE(uint32_t mid, uint32_t ch) override;
+  void WritePCT(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadPCT(uint32_t mid, uint32_t ch) override;
+  void WritePCI(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadPCI(uint32_t mid, uint32_t ch) override;
+  void WritePWT(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadPWT(uint32_t mid, uint32_t ch) override;
+  void WriteDT(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadDT(uint32_t mid, uint32_t ch) override;
+  void WritePSS(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadPSS(uint32_t mid, uint32_t ch) override;
+  void WriteTM(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadTM(uint32_t mid, uint32_t ch) override;
+  void WriteTLT(uint32_t mid, uint32_t data) override;
+  uint32_t ReadTLT(uint32_t mid) override;
+  void WriteSTLT(uint32_t mid, uint32_t ch, uint32_t data) override;
+  uint32_t ReadSTLT(uint32_t mid, uint32_t ch) override;
+  void WriteDSR(uint32_t mid, uint32_t data) override;
+  uint32_t ReadDSR(uint32_t mid) override;
+  void AlignFADC(uint32_t mid) override;
+  void AlignSADC(uint32_t mid) override;
+  void AlignGADC(uint32_t mid) override;
 
   // iadc
-  void WriteDAQMODE(unsigned long mid, unsigned long data) override {}
-  unsigned long ReadDAQMODE(unsigned long mid) override { return 0; }
-  void WriteHV(unsigned long mid, unsigned long ch, float data) override {}
-  float ReadHV(unsigned long mid, unsigned long ch) override { return 0; }
-  float ReadTEMP(unsigned long mid, unsigned long ch) override { return 0; }
-  void AlignIADC(unsigned long mid) override {}
+  void WriteDAQMODE(uint32_t mid, uint32_t data) override {}
+  uint32_t ReadDAQMODE(uint32_t mid) override { return 0; }
+  void WriteHV(uint32_t mid, uint32_t ch, float data) override {}
+  float ReadHV(uint32_t mid, uint32_t ch) override { return 0; }
+  float ReadTEMP(uint32_t mid, uint32_t ch) override { return 0; }
+  void AlignIADC(uint32_t mid) override {}
 
   // tcb
-  void WritePTRIG(unsigned long data) override;
-  unsigned long ReadPTRIG() override;
-  void WriteTRIGENABLE(unsigned long mid, unsigned long data) override;
-  unsigned long ReadTRIGENABLE(unsigned long mid) override;
-  void WriteMTHRFADC(unsigned long data) override;
-  unsigned long ReadMTHRFADC() override;
-  void WritePSCALEFADC(unsigned long data) override;
-  unsigned long ReadPSCALEFADC() override;
-  void WriteMTHRSADCMU(unsigned long data) override {}
-  unsigned long ReadMTHRSADCMU() override { return 0; }
-  void WritePSCALESADCMU(unsigned long data) override {}
-  unsigned long ReadPSCALESADCMU() override { return 0; }
-  void WriteMTHRSADCLS(unsigned long data) override {}
-  unsigned long ReadMTHRSADCLS() override { return 0; }
-  void WritePSCALESADCLS(unsigned long data) override {}
-  unsigned long ReadPSCALESADCLS() override { return 0; }
-  void WriteMTHRIADC(unsigned long data) override {}
-  unsigned long ReadMTHRIADC() override { return 0; }
-  void WritePSCALEIADC(unsigned long data) override {}
-  unsigned long ReadPSCALEIADC() override { return 0; }
-  void WriteEXTOUT(unsigned long data) override {}
-  unsigned long ReadEXTOUT() override { return 0; }
-  void WriteGATEDLY(unsigned long data) override {}
-  unsigned long ReadGATEDLY() override { return 0; }
-  void WriteGATEWIDTH(unsigned long data) override {}
-  unsigned long ReadGATEWIDTH() override { return 0; }
-  void WriteEXTOUTWIDTH(unsigned long data) override {}
-  unsigned long ReadEXTOUTWIDTH() override { return 0; }
+  void WritePTRIG(uint32_t data) override;
+  uint32_t ReadPTRIG() override;
+  void WriteTRIGENABLE(uint32_t mid, uint32_t data) override;
+  uint32_t ReadTRIGENABLE(uint32_t mid) override;
+  void WriteMTHRFADC(uint32_t data) override;
+  uint32_t ReadMTHRFADC() override;
+  void WritePSCALEFADC(uint32_t data) override;
+  uint32_t ReadPSCALEFADC() override;
+  void WriteMTHRSADCMU(uint32_t data) override {}
+  uint32_t ReadMTHRSADCMU() override { return 0; }
+  void WritePSCALESADCMU(uint32_t data) override {}
+  uint32_t ReadPSCALESADCMU() override { return 0; }
+  void WriteMTHRSADCLS(uint32_t data) override {}
+  uint32_t ReadMTHRSADCLS() override { return 0; }
+  void WritePSCALESADCLS(uint32_t data) override {}
+  uint32_t ReadPSCALESADCLS() override { return 0; }
+  void WriteMTHRIADC(uint32_t data) override {}
+  uint32_t ReadMTHRIADC() override { return 0; }
+  void WritePSCALEIADC(uint32_t data) override {}
+  uint32_t ReadPSCALEIADC() override { return 0; }
+  void WriteEXTOUT(uint32_t data) override {}
+  uint32_t ReadEXTOUT() override { return 0; }
+  void WriteGATEDLY(uint32_t data) override {}
+  uint32_t ReadGATEDLY() override { return 0; }
+  void WriteGATEWIDTH(uint32_t data) override {}
+  uint32_t ReadGATEWIDTH() override { return 0; }
+  void WriteEXTOUTWIDTH(uint32_t data) override {}
+  uint32_t ReadEXTOUTWIDTH() override { return 0; }
   void SendTRIG() override;
-  void ReadLNSTAT(unsigned long * data) override;
-  void ReadMIDS(unsigned long * data) override;
-  void AlignDRAM(unsigned long mid) override;
+  void ReadLNSTAT(uint32_t * data) override;
+  void ReadMIDS(uint32_t * data) override;
+  void AlignDRAM(uint32_t mid) override;
 
-  void WriteTRGSWFADC(unsigned long fadc, unsigned long sadcmu, unsigned long sadcls, unsigned long iadc) override {}
-  void WriteTRGSWSADCMU(unsigned long fadc, unsigned long sadcmu, unsigned long sadcls, unsigned long iadc) override {}
-  void WriteTRGSWSADCLS(unsigned long fadc, unsigned long sadcmu, unsigned long sadcls, unsigned long iadc) override {}
-  void WriteTRGSWIADC(unsigned long fadc, unsigned long sadcmu, unsigned long sadcls, unsigned long iadc) override {}
-  unsigned long ReadTRGSWFADC() override { return 0; }
-  unsigned long ReadTRGSWSADCMU() override { return 0; }
-  unsigned long ReadTRGSWSADCLS() override { return 0; }
-  unsigned long ReadTRGSWIADC() override { return 0; }
+  void WriteTRGSWFADC(uint32_t fadc, uint32_t sadcmu, uint32_t sadcls, uint32_t iadc) override {}
+  void WriteTRGSWSADCMU(uint32_t fadc, uint32_t sadcmu, uint32_t sadcls, uint32_t iadc) override {}
+  void WriteTRGSWSADCLS(uint32_t fadc, uint32_t sadcmu, uint32_t sadcls, uint32_t iadc) override {}
+  void WriteTRGSWIADC(uint32_t fadc, uint32_t sadcmu, uint32_t sadcls, uint32_t iadc) override {}
+  uint32_t ReadTRGSWFADC() override { return 0; }
+  uint32_t ReadTRGSWSADCMU() override { return 0; }
+  uint32_t ReadTRGSWSADCLS() override { return 0; }
+  uint32_t ReadTRGSWIADC() override { return 0; }
 
 protected:
   std::string fIPAddress{};
